propagation for counter
