+++ telamon-gen/src/ast/choice/mod.rs
@@ -1,8 +1,10 @@
 pub mod enumeration;
 pub mod integer;
+pub mod counter;
 
 pub use self::enumeration::EnumDef;
 pub use self::integer::IntegerDef;
+pub use self::counter::CounterDef;
 pub use super::*;
 
 #[derive(Debug, PartialEq, Clone)]
@@ -34,9 +36,9 @@ impl ChoiceDef {
         tc: &mut TypingContext,
     ) -> Result<(), TypeError> {
         match self {
-            ChoiceDef::CounterDef(_) => Ok(()),
-            ChoiceDef::IntegerDef(integer_def) => integer_def.define(context, tc),
-            ChoiceDef::EnumDef(enum_def) => enum_def.define(context, tc),
+            ChoiceDef::CounterDef(def) => def.define(context, tc),
+            ChoiceDef::IntegerDef(def) => def.define(context, tc),
+            ChoiceDef::EnumDef(def) => def.define(context, tc),
         }
     }
 }
